% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nspn500}
\alias{nspn500}
\title{NSPN500 Cortical Atlas}
\usage{
nspn500()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Neuroscience in Psychiatry Network (NSPN) 500
cortical parcellation. Contains 2D polygon geometry for
\code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
nspn500()
plot(nspn500())
}
\references{
Whitaker KJ, Vertes PE, Romero-Garcia R, et al. (2016).
Adolescence is associated with genomically patterned consolidation
of the hubs of the human brain connectome.
\emph{Proceedings of the National Academy of Sciences}, 113(32):9105-9110.
\doi{10.1073/pnas.1601745113}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
