% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nspn500}
\alias{nspn500}
\title{NSPN500 Cortical Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
nspn500
}
\description{
NSPN500 Cortical Atlas
}
\keyword{datasets}
